+++ catalyst/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "21.02rc0"
+__version__ = "21.02rc1"
