+++ catalyst/core/__init__.py
@@ -1,10 +1,9 @@
 # flake8: noqa
 # import order:
-# experiment
 # runner
 # callback
 
-from catalyst.core.experiment import IExperiment
+# from catalyst.core.experiment import IExperiment
 from catalyst.core.runner import IRunner, IStageBasedRunner, RunnerException
 from catalyst.core.callback import (
     Callback,
