+++ catalyst/core/runner.py
@@ -1,28 +1,30 @@
-from typing import Any, Dict, Mapping, Tuple
+from typing import Any, Dict, Iterable, Mapping, Tuple
 from abc import ABC, abstractmethod
-from collections import defaultdict
+from collections import defaultdict, OrderedDict
 from functools import lru_cache, partial
 import logging
 
 import torch
 import torch.distributed
 import torch.multiprocessing
-from torch.utils.data import DataLoader, DistributedSampler
+from torch.utils.data import DataLoader, Dataset, DistributedSampler
 
-from catalyst.core.callback import Callback, CallbackScope, ICallback
+from catalyst.core.callback import Callback, ICallback
 from catalyst.core.engine import IEngine
-from catalyst.core.experiment import IExperiment
 from catalyst.core.functional import filter_callbacks_by_node, sort_callbacks_by_order
 from catalyst.core.logger import ILogger
 from catalyst.core.trial import ITrial
 from catalyst.engines.distributed import DistributedDataParallelEngine
 from catalyst.typing import (
+    Criterion,
     Device,
     Model,
+    Optimizer,
     RunnerCriterion,
     RunnerModel,
     RunnerOptimizer,
     RunnerScheduler,
+    Scheduler,
 )
 from catalyst.utils.loaders import validate_loaders
 from catalyst.utils.misc import maybe_recursive_call, set_global_seed
@@ -77,7 +79,7 @@ class IRunner(ICallback, ILogger, ABC):
         self.model: RunnerModel = model
         # self._device = device
         self.engine: IEngine = engine
-        self.experiment: IExperiment = None
+        # self.experiment: IExperiment = None
         self.trial: ITrial = None
         # the data
         self.loaders: Dict[str, DataLoader] = None
@@ -99,7 +101,7 @@ class IRunner(ICallback, ILogger, ABC):
         self.epoch_metrics: EPOCH_METRICS = defaultdict(None)
 
         # experiment info
-        self.experiment_key: str = None
+        self.run_key: str = None
         self.global_epoch_step: int = 0
         self.global_batch_step: int = 0
         self.global_sample_step: int = 0
@@ -138,13 +140,283 @@ class IRunner(ICallback, ILogger, ABC):
         """Returns the runner's device instance."""
         return self.engine.device
 
+    @property
+    def seed(self) -> int:
+        """Experiment's seed for reproducibility."""
+        return 42
+
+    @property
+    def name(self) -> str:
+        # @TODO: return the hash from the IRunner instance
+        return "IRunner"
+
+    @property
+    def hparams(self) -> OrderedDict:
+        """
+        Returns hyper-parameters for current run.
+
+        Example::
+            >>> runner.hparams
+            OrderedDict([('optimizer', 'Adam'),
+             ('lr', 0.02),
+             ('betas', (0.9, 0.999)),
+             ('eps', 1e-08),
+             ('weight_decay', 0),
+             ('amsgrad', False),
+             ('train_batch_size', 32)])
+        """
+        return {}
+
+    @property
+    @abstractmethod
+    def stages(self) -> Iterable[str]:
+        """Experiment's stage names.
+
+        Example::
+
+            >>> runner.stages
+            ["pretraining", "finetuning"]
+        """
+        pass
+
+    def get_stage_len(self, stage: str) -> int:
+        return 1
+
+    def get_trial(self) -> ITrial:
+        return None
+
+    @abstractmethod
+    def get_engine(self) -> IEngine:
+        return None
+
+    def get_loggers(self) -> Dict[str, ILogger]:
+        return {}
+
+    def get_transforms(self, stage: str = None, epoch: int = None, dataset: str = None, **kwargs):
+        """Returns the data transforms for a given stage and dataset.
+
+        Args:
+            stage: stage name of interest,
+                like "pretrain" / "train" / "finetune" / etc
+            dataset: dataset name of interest,
+                like "train" / "valid" / "infer"
+
+        .. note::
+            For datasets/loaders naming please follow
+            :py:mod:`catalyst.core.runner` documentation.
+
+        Returns:  # noqa: DAR202
+            Data transformations to use for specified dataset.
+
+        """
+        raise NotImplementedError
+
+    def get_datasets(self, stage: str, epoch: int = None, **kwargs) -> "OrderedDict[str, Dataset]":
+        """Returns the datasets for a given stage and epoch.  # noqa: DAR401
+
+        .. note::
+            For Deep Learning cases you have the same dataset
+            during whole stage.
+
+            For Reinforcement Learning it's common to change the dataset
+            (experiment) every training epoch.
+
+        Args:
+            stage: stage name of interest,
+                like "pretrain" / "train" / "finetune" / etc
+            epoch: epoch index
+            **kwargs: additional parameters to use during
+                dataset creation
+
+        Returns:  # noqa: DAR202
+            OrderedDict[str, Dataset]: Ordered dictionary
+                with datasets for current stage and epoch.
+
+        .. note::
+            We need ordered dictionary to guarantee the correct dataflow
+            and order of our training datasets.
+            For example, to run train loader before validation one :)
+
+        Example::
+
+            >>> experiment.get_datasets(
+            >>>     stage="training",
+            >>>     in_csv_train="path/to/train/csv",
+            >>>     in_csv_valid="path/to/valid/csv",
+            >>> )
+            OrderedDict({
+                "train": CsvDataset(in_csv=in_csv_train, ...),
+                "valid": CsvDataset(in_csv=in_csv_valid, ...),
+            })
+
+
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    def get_loaders(self, stage: str, epoch: int = None) -> "OrderedDict[str, DataLoader]":
+        """Returns the loaders for a given stage.  # noqa: DAR401
+
+        .. note::
+            Wrapper for
+            :py:mod:`catalyst.core.experiment.IExperiment.get_datasets`.
+            For most of your experiments you need to rewrite `get_datasets`
+            method only.
+
+        Args:
+            stage: stage name of interest,
+                like "pretrain" / "train" / "finetune" / etc
+            epoch: epoch index
+
+        Returns:  # noqa: DAR202
+            OrderedDict[str, DataLoader]: Ordered dictionary
+                with loaders for current stage and epoch.
+
+        """
+        pass
+
+    @abstractmethod
+    def get_model(self, stage: str, epoch: int = None) -> Model:
+        """Returns the model for a given stage and epoch.
+
+        Example::
+
+            # suppose we have typical MNIST model, like
+            # nn.Sequential(nn.Linear(28*28, 128), nn.Linear(128, 10))
+            >>> runner.get_model(stage="train")
+            Sequential(
+             : Linear(in_features=784, out_features=128, bias=True)
+             : Linear(in_features=128, out_features=10, bias=True)
+            )
+
+        Args:
+            stage: stage name of interest
+                like "pretrain" / "train" / "finetune" / etc
+
+        Returns:  # noqa: DAR202
+            Model: model for a given stage.
+        """
+        pass
+
+    def get_criterion(self, stage: str, epoch: int = None) -> Criterion:
+        """Returns the criterion for a given stage and epoch.
+
+        Example::
+
+            # for typical classification task
+            >>> runner.get_criterion(stage="train")
+            nn.CrossEntropyLoss()
+
+        Args:
+            stage: stage name of interest
+                like "pretrain" / "train" / "finetune" / etc
+
+        Returns:  # noqa: DAR202
+            Criterion: criterion for a given stage.
+        """
+        return None
+
+    def get_optimizer(self, model: Model, stage: str, epoch: int = None) -> Optimizer:
+        """Returns the optimizer for a given stage and model.
+
+        Example::
+
+            >>> runner.get_optimizer(model=model, stage="train")
+            torch.optim.Adam(model.parameters())
+
+        Args:
+            stage: stage name of interest
+                like "pretrain" / "train" / "finetune" / etc
+            model: model to optimize with stage optimizer
+
+        Returns:  # noqa: DAR202
+            Optimizer: optimizer for a given stage and model.
+        """
+        return None
+
+    def get_scheduler(self, optimizer: Optimizer, stage: str, epoch: int = None) -> Scheduler:
+        """Returns the scheduler for a given stage and optimizer.
+
+        Example::
+            >>> runner.get_scheduler(stage="training", optimizer=optimizer)
+            torch.optim.lr_scheduler.StepLR(optimizer)
+
+        Args:
+            stage: stage name of interest
+                like "pretrain" / "train" / "finetune" / etc
+            optimizer: optimizer to schedule with stage scheduler
+
+        Returns:  # noqa: DAR202
+            Scheduler: scheduler for a given stage and optimizer.
+        """
+        return None
+
+    def _get_model(self) -> Model:
+        self.model = self.get_model(stage=self.stage_key, epoch=self.stage_epoch_step)
+        return self.model
+
+    def _get_criterion(self) -> Criterion:
+        self.criterion = self.get_criterion(stage=self.stage_key, epoch=self.stage_epoch_step)
+        return self.criterion
+
+    def _get_optimizer(self) -> Optimizer:
+        assert self.model is not None, "You need to setup model first"
+        self.optimizer = self.get_optimizer(
+            stage=self.stage_key, epoch=self.stage_epoch_step, model=self.model
+        )
+        return self.optimizer
+
+    def _get_scheduler(self) -> Scheduler:
+        assert self.optimizer is not None, "You need to setup optimizer first"
+        self.scheduler = self.get_scheduler(
+            stage=self.stage_key, epoch=self.stage_epoch_step, optimizer=self.optimizer
+        )
+        return self.scheduler
+
+    def get_callbacks(self, stage: str, epoch: int = None) -> "OrderedDict[str, ICallback]":
+        """Returns callbacks for a given stage.
+
+        .. note::
+            To learn more about Catalyst Callbacks mechanism, please follow
+            :py:mod:`catalyst.core.callback.Callback` documentation.
+
+        .. note::
+            We need ordered dictionary to guarantee the correct dataflow
+            and order of metrics optimization.
+            For example, to compute loss before optimization ;)
+
+        Args:
+            stage: stage name of interest
+                like "pretrain" / "train" / "finetune" / etc
+
+        Returns:  # noqa: DAR202
+            OrderedDict[str, Callback]: Ordered dictionary  # noqa: DAR202
+            with callbacks for current stage.
+
+        .. note::
+            To learn more about Catalyst Core concepts, please check out
+
+                - :py:mod:`catalyst.core.experiment.IExperiment`
+                - :py:mod:`catalyst.core.runner.IRunner`
+                - :py:mod:`catalyst.core.callback.Callback`
+
+        Args:
+            stage: stage name of interest,
+                like "pretrain" / "train" / "finetune" / etc
+
+        Returns:
+            OrderedDict[str, Callback]: Ordered dictionary
+                with callbacks for current stage.
+        """
+        return {}
+
     def log_metrics(self, *args, **kwargs) -> None:
         for logger in self.loggers.values():
             logger.log_metrics(
                 *args,
                 **kwargs,
                 # experiment info
-                experiment_key=self.experiment_key,
+                experiment_key=self.run_key,
                 global_sample_step=self.global_sample_step,
                 global_batch_step=self.global_batch_step,
                 global_epoch_step=self.global_epoch_step,
@@ -168,7 +440,7 @@ class IRunner(ICallback, ILogger, ABC):
                 *args,
                 **kwargs,
                 # experiment info
-                experiment_key=self.experiment_key,
+                experiment_key=self.run_key,
                 global_sample_step=self.global_sample_step,
                 global_batch_step=self.global_batch_step,
                 global_epoch_step=self.global_epoch_step,
@@ -192,7 +464,7 @@ class IRunner(ICallback, ILogger, ABC):
                 *args,
                 **kwargs,
                 # experiment info
-                experiment_key=self.experiment_key,
+                experiment_key=self.run_key,
             )
 
     def flush_log(self) -> None:
@@ -204,8 +476,7 @@ class IRunner(ICallback, ILogger, ABC):
             logger.close_log()
 
     def on_experiment_start(self, runner: "IRunner"):
-        assert self.experiment is not None
-        self.experiment_key = self.experiment.name
+        self.run_key = self.name
         self.global_epoch_step: int = 0
         self.global_batch_step: int = 0
         self.global_sample_step: int = 0
@@ -213,16 +484,15 @@ class IRunner(ICallback, ILogger, ABC):
         self.need_early_stop: bool = False
         self.need_exception_reraise: bool = True
 
-        self.trial = self.experiment.get_trial()
-        self.engine = self.experiment.get_engine()
-        self.loggers = self.experiment.get_loggers()
-        self.log_hparams(hparams=self.experiment.hparams)
+        self.trial = self.get_trial()
+        self.engine = self.get_engine()
+        self.loggers = self.get_loggers()
+        self.log_hparams(hparams=self.hparams)
 
     def on_stage_start(self, runner: "IRunner"):
         assert self.stage_key is not None
-        stage_params = self.experiment.get_stage_params(self.stage_key)
         self.is_infer_stage: bool = self.stage_key.startswith("infer")
-        self.stage_epoch_len = stage_params["num_epochs"]
+        self.stage_epoch_len = self.get_stage_len(stage=self.stage_key)
         self.stage_epoch_step: int = 0
         self.stage_batch_step: int = 0
         self.stage_sample_step: int = 0
@@ -358,7 +628,7 @@ class IRunner(ICallback, ILogger, ABC):
 
     def _run_experiment(self) -> None:
         self._run_event("on_experiment_start")
-        for self.stage_key in self.experiment.stages:
+        for self.stage_key in self.stages:
             if self.engine.rank < 0:
                 # single-device branch (cpu, gpu, dp)
                 self._run_stage()
@@ -371,7 +641,7 @@ class IRunner(ICallback, ILogger, ABC):
                 # raise NotImplementedError()
         self._run_event("on_experiment_end")
 
-    def run(self, experiment: IExperiment = None) -> "IRunner":
+    def run(self) -> "IRunner":
         """Runs the experiment.
 
         Args:
@@ -380,7 +650,7 @@ class IRunner(ICallback, ILogger, ABC):
         Returns:
             self, `IRunner` instance after the experiment
         """
-        self.experiment = experiment or self.experiment
+        # self.experiment = experiment or self.experiment
         try:
             self._run_experiment()
         except (Exception, KeyboardInterrupt) as ex:
@@ -393,35 +663,13 @@ class IStageBasedRunner(IRunner):
     """Runner abstraction that suppose to have constant datasources per stage."""
 
     def _setup_loaders(self) -> None:
-        loaders = self.experiment.get_loaders(stage=self.stage_key)
+        set_global_seed(self.seed + self.engine.rank + self.global_epoch_step)
+        loaders = self.get_loaders(stage=self.stage_key)
         loaders = validate_loaders(loaders)
         self.loaders = loaders
 
-    def _get_model(self):
-        stage_params = self.experiment.get_stage_params(stage=self.stage_key)
-        migrate_model_from_previous_stage = stage_params.get(
-            "migrate_model_from_previous_stage", True
-        )
-        # some custom logic is possible here
-        if self.model is not None and migrate_model_from_previous_stage:
-            return self.model
-        else:
-            return self.experiment.get_model(stage=self.stage_key)
-
-    def _get_criterion(self):
-        # some custom logic is possible here
-        return self.experiment.get_criterion(stage=self.stage_key)
-
-    def _get_optimizer(self, model):
-        # some custom logic is possible here
-        return self.experiment.get_optimizer(stage=self.stage_key, model=model)
-
-    def _get_sheduler(self, optimizer):
-        # some custom logic is possible here
-        return self.experiment.get_scheduler(stage=self.stage_key, optimizer=optimizer)
-
     def _setup_components(self) -> None:
-        # @TODO: should we move engine logic here?
+        set_global_seed(self.seed + self.engine.rank + self.global_epoch_step)
         (
             self.model,
             self.criterion,
@@ -431,36 +679,23 @@ class IStageBasedRunner(IRunner):
             model_fn=self._get_model,
             criterion_fn=self._get_criterion,
             optimizer_fn=self._get_optimizer,
-            scheduler_fn=self._get_sheduler,
+            scheduler_fn=self._get_scheduler,
         )
 
     def _setup_callbacks(self):
-        stage_params = self.experiment.get_stage_params(stage=self.stage_key)
-        migrate_callbacks_from_previous_stage = stage_params.get(
-            "migrate_callbacks_from_previous_stage", True
-        )
-        # seed fix during callbacks creation for reproducibility
-        set_global_seed(self.experiment.seed + self.engine.rank + self.global_epoch_step)
-        callbacks = self.experiment.get_callbacks(self.stage_key)
+        set_global_seed(self.seed + self.engine.rank + self.global_epoch_step)
+        callbacks = self.get_callbacks(self.stage_key)
         callbacks = filter_callbacks_by_node(callbacks)
-        callbacks = sort_callbacks_by_order(callbacks)
-        if migrate_callbacks_from_previous_stage:
-            for key, value in self.callbacks.items():
-                if value.scope == CallbackScope.experiment:
-                    callbacks[key] = value
         callbacks = sort_callbacks_by_order(callbacks)
         self.callbacks = callbacks
 
     def on_stage_start(self, runner: "IRunner") -> None:
         super().on_stage_start(runner)
         # the data
-        set_global_seed(self.experiment.seed + self.engine.rank + self.global_epoch_step)
         self._setup_loaders()
         # the components
-        set_global_seed(self.experiment.seed + self.engine.rank + self.global_epoch_step)
         self._setup_components()
         # the callbacks
-        set_global_seed(self.experiment.seed + self.engine.rank + self.global_epoch_step)
         self._setup_callbacks()
 
     def on_epoch_start(self, runner: "IRunner"):
@@ -469,13 +704,14 @@ class IStageBasedRunner(IRunner):
         for loader_key, loader in self.loaders.items():
             if len(loader) == 0:
                 raise RunnerException(f"DataLoader with name {loader_key} is empty.")
+        set_global_seed(self.seed + self.engine.rank + self.global_epoch_step)
 
     def on_loader_start(self, runner: "IRunner"):
         super().on_loader_start(runner)
         self.loader_batch_len = len(self.loader)
         if self.loader_batch_len == 0:
             raise NotImplementedError(f"DataLoader with name {self.loader_key} is empty.")
-        set_global_seed(self.experiment.seed + self.engine.rank + self.global_epoch_step)
+        set_global_seed(self.seed + self.engine.rank + self.global_epoch_step)
 
     def on_stage_end(self, runner: "IRunner") -> None:
         # clean process if DDP training or do nothing
