+++ catalyst/dl/__init__.py
@@ -2,7 +2,8 @@
 
 from catalyst.core import *
 from catalyst.engines import *
-from catalyst.experiments import *
+
+# from catalyst.experiments import *
 from catalyst.runners import *
 from catalyst.callbacks import *
 from catalyst.loggers import *
