+++ catalyst/experiments/__init__.py
@@ -2,7 +2,7 @@
 
 # from catalyst.experiments.misc import SingleStageExperiment
 
-from catalyst.experiments.experiment import Experiment
+# from catalyst.experiments.experiment import Experiment
 
 # from catalyst.experiments.auto import AutoCallbackExperiment
 # from catalyst.experiments.config import ConfigExperiment
