+++ catalyst/experiments/tests/test_core.py
@@ -56,9 +56,9 @@ def test_defaults():
         "minimize_metric": True,
         "checkpoint_data": {},
     }
-    assert exp.get_model("") == model
-    assert exp.get_criterion("") is None
-    assert exp.get_optimizer("", model) is None
-    assert exp.get_scheduler("") is None
+    assert exp._get_model("") == model
+    assert exp._get_criterion("") is None
+    assert exp._get_optimizer("", model) is None
+    assert exp._get_scheduler("") is None
     _test_callbacks(test_callbacks, exp)
     assert exp.get_loaders("") == loaders
