+++ catalyst/utils/quantization.py
@@ -107,7 +107,7 @@ def quantize_model_from_checkpoint(
     if stage is None:
         stage = list(experiment.stages)[0]
 
-    model = experiment.get_model(stage)
+    model = experiment._get_model(stage)
     checkpoint = load_checkpoint(checkpoint_path)
     unpack_checkpoint(checkpoint, model=model)
